Refactor src-tauri: Split backend.rs and main.rs into modular files
