+++ apps/tauri/src-tauri/build.rs
@@ -1,5 +1,5 @@
 fn main() {
-	println!("cargo:rerun-if-changed=tauri.conf.json");
-	println!("cargo:rerun-if-changed=../dist");
-	tauri_build::build()
+    println!("cargo:rerun-if-changed=tauri.conf.json");
+    println!("cargo:rerun-if-changed=../dist");
+    tauri_build::build()
 }
