+++ apps/tauri/src-tauri/src/commands/mod.rs
@@ -1,0 +1,3 @@
+pub mod fs;
+pub mod media;
+pub mod utils;
